requirements.txt
absl-py==2.3.1
accelerate==1.8.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.9.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
boto3==1.39.14
botocore==1.39.14
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.59.0
datasets==3.6.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
Django==5.2.4
dnspython==2.7.0
docker==7.1.0
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.0
fastapi==0.116.0
ffmpy==0.6.0
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.5
fpdf==1.7.2
fpdf2==2.8.3
from-root==1.3.0
frontend==0.0.3
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-documentai==3.5.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==5.35.0
gradio_client==1.10.4
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
groovy==0.1.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.8
libclang==18.1.1
looseversion==1.3.0
lxml==6.0.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mlflow==3.1.1
mlflow-skinny==3.1.1
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mypy-boto3-s3==1.39.5
namex==0.1.0
narwhals==1.48.1
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.1.3
ollama==0.5.1
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.22.1
onnxslim==0.1.61
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psutil==7.0.0
pure_eval==0.2.3
puremagic==1.30
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
pymongo==4.13.2
PyMuPDF==1.26.3
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pywin32==311
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.0.0
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
ruff==0.12.2
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
selenium==4.34.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.20.1
simsimd==6.5.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tokenizers==0.21.2
tomlkit==0.13.3
tools==1.0.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.53.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.163
ultralytics-thop==2.0.14
urllib3==2.5.0
uv==0.8.4
uvicorn==0.35.0
waitress==3.0.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
